Calculate forces in absolute coordinates
Previously, we were calculating them in relative coordinates then
converting the net force, since that seemed simpler. That no longer
appears to be the case, especially given that we kept having to take the
inverse of the orientation to get relative coordinates from absolute.
This way all conversions go the same way.